backend/requirements.txt
joblib==1.3.2
Flask-Cors==4.0.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
openai==0.27.9
python-dotenv==0.19.1
psycopg2-binary==2.9.7
pydantic
pandas
python-dotenv
faiss-cpu
tenacity
pytest
tiktoken